Implemented DbManager.getRevisionData() and changed RevisionInfo to support larger numbers. Table structure changed to store diffs as text instead of Blob
git-svn-id: http://fbms.googlecode.com/svn/trunk@106 c375ce47-9a4c-63b9-7dad-5ffa0e8ccef7

+++ main/src/cmpt370/fbms/RevisionInfo.java
@@ -2,9 +2,9 @@ package cmpt370.fbms;
 
 public class RevisionInfo
 {
-	public int id;
+	public long id;
 	public String path;
 	public String diff;
-	public int delta;
-	public int time;
+	public long delta;
+	public long time;
 }
